Split routes/appointments.js along its seams
3,500 lines into app/routes/appointments/: index (context middleware,
module registration, dynamic template fallback last), lifecycle,
participant-details, previous-mammograms, symptoms, medical-information,
medical-history, imaging-automatic, imaging-manual, special-appointments,
appointment-notes, cancel-reschedule, plus shared.js for the
cross-module captureSessionEndTime helper. Route handlers are moved
verbatim; only require blocks are regenerated. Route inventory verified
identical (50 routes) before and after.
